# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
  api.ApiResponse:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
...
  api.loginRequest:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  api.loginResponse:
    properties:
      access_token:
        type: string
      expires_in:
        type: integer
      refresh_token:
        type: string
    type: object
...
  api.refreshRequest:
    properties:
      refresh_token:
        type: string
    required:
    - refresh_token
    type: object
  api.refreshResponse:
    properties:
      access_token:
        type: string
      expires_in:
        type: integer
    type: object
  api.registerRequest:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  api.registerResponse:
    properties:
      user_id:
        type: integer
      username:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: GoIM 即时通讯系统 API — 支持私聊、群聊、朋友圈动态。
  title: GoIM API
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: 使用用户名和密码登录，返回访问令牌和刷新令牌
      parameters:
      - description: 登录信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.loginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 登录成功
          schema:
            allOf:
            - $ref: '#/definitions/api.ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/api.loginResponse'
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/api.ApiResponse'
        "401":
          description: 用户名或密码错误
          schema:
            $ref: '#/definitions/api.ApiResponse'
      summary: 用户登录
      tags:
      - 认证
  /auth/refresh:
    post:
      consumes:
      - application/json
      description: 使用刷新令牌获取新的访问令牌
      parameters:
      - description: 刷新令牌
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.refreshRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 刷新成功
          schema:
            allOf:
            - $ref: '#/definitions/api.ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/api.refreshResponse'
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/api.ApiResponse'
        "401":
          description: 令牌无效或已过期
          schema:
            $ref: '#/definitions/api.ApiResponse'
      summary: 刷新令牌
      tags:
      - 认证
  /auth/register:
    post:
      consumes:
      - application/json
      description: 使用用户名和密码注册新用户
      parameters:
      - description: 注册信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.registerRequest'
      produces:
      - application/json
      responses:
        "201":
          description: 注册成功
          schema:
            allOf:
            - $ref: '#/definitions/api.ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/api.registerResponse'
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/api.ApiResponse'
        "409":
          description: 用户名已存在
          schema:
            $ref: '#/definitions/api.ApiResponse'
      summary: 用户注册
      tags:
      - 认证
  /avatar/{userID}:
    get:
      description: 返回用户头像图片。如果用户未设置头像，自动生成包含用户名首字母的 SVG 占位头像
      parameters:
      - description: 用户ID
        format: int64
        in: path
        name: userID
        required: true
        type: integer
      - description: 用户名（用于生成默认头像首字母）
        in: query
        name: name
        type: string
      produces:
      - image/svg+xml
      responses:
        "200":
          description: 头像图片（SVG或重定向到图片URL）
          schema:
            type: string
      summary: 获取用户头像
      tags:
      - 用户
...
securityDefinitions:
  BearerAuth:
    description: 在 Authorization 头中携带 JWT 访问令牌，格式：Bearer <token>
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"